# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: WoloApp
  version: "1.0.0"
  description: >-
    The WoloApp API facilitates the management of events and user participation
    through Create, Read, Update, and Delete operations. It allows users to
    interact with the following key functionalities such as exploring events
    through different filters or joining them.
  termsOfService: >-
    https://opensource.guide/legal/?fbclid=IwAR0NUY2SAm3x5HHkhxOmb6piB-WETsmqeifvIB5_pBrHWFhQu31oYBl7WJs
  contact:
    name: Agata Dobrzyniewicz
    email: s24305@pjwstk.edu.pl
servers:
  - url: http://localhost:8080
    description: local
paths:
...
  /organisations/{id}:
    get:
      tags:
        - Organisations
      summary: Get organisation by id
      description: Organisation must exist
      operationId: getOrganisationById
      parameters:
        - name: id
          in: path
          description: Organisation id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganisationResponseDto'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /organisations/{id}/events:
    get:
      tags:
        - Organisations
      summary: Get all events from the selected organisation
      operationId: getEventsByOrganisation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDto'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /organisations/events/current/{id}:
    get:
      tags:
        - Organisations
      summary: Get all current events from the selected organisation
      operationId: getCurrenEventsByOrganisation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDto'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /organisations/events/past/{id}:
    get:
      tags:
        - Organisations
      summary: Get all past events from the selected organisation
      operationId: getPastEventsByOrganisation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponseDto'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /organisations/add:
    post:
      tags:
        - Organisations
      summary: Adding new organisation
      description: id = null
      operationId: addOrganisation
      parameters:
        - name: language
          in: query
          required: true
          schema:
            type: string
          example: "PL"
        - name: organisationDto
          in: query
          description: Organisation object to create
          schema:
            $ref: '#/components/schemas/OrganisationRequestDto'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganisationRequestDto'
        required: true
      responses:
        '201':
          $ref: '#/components/responses/Created'
  /organisations/approve:
    post:
      tags:
        - Organisations
      summary: Approve organisation (by admin)
      description: Organisation must exist
      operationId: approveOrganisation
      parameters:
        - name: id
          in: query
          description: Organisation id to approve
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
  /organisations/disapprove:
    post:
      tags:
        - Organisations
      summary: Disapprove organisation (by admin)
      description: Organisation must exist
      operationId: disapproveOrganisation
      parameters:
        - name: id
          in: query
          description: Organisation id to disapprove
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /organisations/{id}/edit:
    put:
      tags:
        - Organisations
      summary: Edit organisation
      description: Organisation must exist
      operationId: editOrganisation
      parameters:
        - name: organisation
          in: query
          description: Organisation object with changes
          schema:
            $ref: '#/components/schemas/OrganisationRequestDto'
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: language
          in: query
          required: true
          schema:
            type: integer
          example: "PL"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganisationRequestDto'
        required: true
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
  /reports/all:
    get:
      tags:
        - Reports
      summary: Get all reports from one event
      operationId: getAllReportsByEventId
      parameters:
        - name: eventId
          in: query
          description: Event Id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportResponceDto'
  /reports/public/{id}:
    get:
      tags:
        - Reports
      summary: Get public report by event id
      description: Report and event must exist
      operationId: getPublicReportByEventId
      parameters:
        - name: id
          in: path
          description: Report id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponceDto'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /reports/add:
    post:
      tags:
        - Reports
      summary: Adding report
      description: id = null
      operationId: addReport
      parameters:
        - name: language
          in: query
          description: Language of ReportDto object
          required: true
          schema:
            type: string
          example: "PL"
        - name: report
          in: query
          description: Report object to create
          schema:
            $ref: '#/components/schemas/ReportRequestDto'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequestDto'
        required: true
      responses:
        '201':
          $ref: '#/components/responses/Created'
  /reports/edit:
    put:
      tags:
        - Reports
      summary: Edit report
      description: Report must exist
      operationId: editReport
      parameters:
        - name: language
          in: query
          description: Language of ReportDto object
          required: true
          schema:
            type: string
          example: "PL"
        - name: report
          in: query
          description: Report object with changes
          schema:
            $ref: '#/components/schemas/ReportRequestDto'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequestDto'
        required: true
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
  /reports/delete/{id}:
    delete:
      tags:
        - Reports
      summary: Delete report
      operationId: deleteReport
      parameters:
        - name: id
          in: path
          description: Report id
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '404':
          $ref: '#/components/responses/NotFoundResponse'
...
components:
  responses:
    OK:
      description: OK
      content:
        application/hal+json:
          schema:
            type: string
            example: Ok
            enum:
              - 200 OK
    Created:
      description: Created
      content:
        application/hal+json:
          schema:
            type: string
            example: Created
            enum:
              - 201 CREATED
    NoContentResponse:
      description: No content
      content:
        application/hal+json:
          schema:
            type: string
            example: No Content
            enum:
              - 204 NO_CONTENT
    BadRequestResponse:
      description: Bad Request
      content:
        application/hal+json:
          schema:
            type: string
            example: Bad Request
            enum:
              - 400 BAD_REQUEST
    NotFoundResponse:
      description: Not Found
      content:
        application/hal+json:
          schema:
            type: string
            example: Not Found
            enum:
              - 404 NOT_FOUND
...
  schemas:
...
    EventResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nameEN:
          type: string
          example: Animal shelter assistance
        namePL:
          type: string
        nameUA:
          type: string
        nameRU:
          type: string
        organisation:
          type: string
          example: 'Compassion Paws: Animal Shelter Support Initiative'
        isPeselVerificationRequired:
          type: boolean
        categories:
          type: array
          items:
            type: string
            example: Health
        city:
          type: string
          example: New York City
        date:
          type: string
          format: date
        imageUrl:
          type: string
          example: https://example.com/image.jpg
        alt:
          type: string
        shifts:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/ShiftResponseDto'
    OrganisationRequestDto:
      required:
        - name
        - description
        - email
        - phoneNumber
        - street
        - homeNum
        - districtId
        - cityId
        - moderatorId
      type: object
      properties:
        name:
          maxLength: 250
          minLength: 0
          type: string
          example: Animal Helpers
        description:
          type: string
          example: >-
            Are you passionate about making a positive impact on the lives of
            animals in need? Animal Helpers is

            dedicated to providing compassionate care and support for animals at
            our local animal shelter.

            As part of our organization, you'll have the opportunity to play a
            vital role in ensuring the well-being

            and happiness of animals awaiting their forever homes.


            Our Mission:
                - Provide shelter and care for animals in need.
                - Facilitate adoptions and find loving homes for our animals.
                - Educate the community about responsible pet ownership and animal welfare.

            Volunteer Opportunities:
                - Animal Care: Assist in feeding, grooming, and providing basic care for shelter animals.
                - Adoption Assistance: Interact with potential adopters and facilitate the adoption process.
                - Community Outreach: Participate in events and educational programs to promote animal welfare.

            Join us today and make a difference in the lives of animals in need!
        email:
          maxLength: 50
          minLength: 0
          type: string
        phoneNumber:
          maxLength: 9
          minLength: 9
          pattern: '[0-9]{9}'
          type: string
        street:
          maxLength: 50
          minLength: 0
          type: string
        homeNum:
          maxLength: 10
          minLength: 0
          type: string
        districtId:
          type: integer
          format: int64
          example: 1
        cityId:
          type: integer
          format: int64
          example: 1
        moderatorId:
          type: integer
          format: int64
          example: 1
        logoUrl:
          maxLength: 255
          minLength: 0
          type: string
          example: https://example.com/image.jpg
...
    OrganisationResponseDto:
      type: object
      properties:
        name:
          maxLength: 250
          minLength: 0
          type: string
          example: Animal Helpers
        descriptionPL:
          type: string
          example: >-
            Are you passionate about making a positive impact on the lives of
            animals in need? Animal Helpers is

            dedicated to providing compassionate care and support for animals at
            our local animal shelter.

            As part of our organization, you'll have the opportunity to play a
            vital role in ensuring the well-being

            and happiness of animals awaiting their forever homes.


            Our Mission:
                - Provide shelter and care for animals in need.
                - Facilitate adoptions and find loving homes for our animals.
                - Educate the community about responsible pet ownership and animal welfare.

            Volunteer Opportunities:
                - Animal Care: Assist in feeding, grooming, and providing basic care for shelter animals.
                - Adoption Assistance: Interact with potential adopters and facilitate the adoption process.
                - Community Outreach: Participate in events and educational programs to promote animal welfare.

            Join us today and make a difference in the lives of animals in need!
        descriptionEN:
          type: string
        descriptionUA:
          type: string
        descriptionRU:
          type: string
        email:
          maxLength: 50
          minLength: 0
          type: string
          example: example@mail.com
        phoneNumber:
          maxLength: 9
          minLength: 9
          type: string
        street:
          maxLength: 50
          minLength: 0
          type: string
          example: Broadway
        homeNum:
          maxLength: 10
          minLength: 0
          type: string
          example: 123
        logoUrl:
          maxLength: 255
          minLength: 0
          type: string
          example: https://example.com/image.jpg
    ReportRequestDto:
      required:
        - report
        - event
        - published
      properties:
        id:
          type: integer
          format: int64
          example: 1
        report:
          type: string
        published:
          type: boolean
        event:
          type: integer
          format: int64
          example: 1
    ReportResponceDto:
      properties:
        id:
          type: integer
          format: int64
          example: 1
        reportPL:
          type: string
        reportEN:
          type: string
        reportUA:
          type: string
        reportRU:
          type: string
        published:
          type: boolean
        event:
          type: integer
          format: int64
          example: 1
...
    ShiftResponseDto:
      type: object
      properties:
        shiftId:
          type: integer
          format: int64
          example: 1
        eventId:
          type: integer
          format: int64
          example: 1
        eventNamePL:
          type: string
          example: Pomocnicy zwierząt
        eventNameEN:
          type: string
          example: Animal Helpers
        eventNameUA:
          type: string
          example: Помічники тварин
        eventNameRU:
          type: string
          example: Помощники для животных
        date:
          type: string
          format: date
        startTime:
          $ref: '#/components/schemas/LocalTime'
        endTime:
          $ref: '#/components/schemas/LocalTime'
        capacity:
          type: integer
          format: int32
          example: 5
        isLeaderRequired:
          type: boolean
        requiredMinAge:
          type: integer
          format: int32
          example: 12
        registeredUsers:
          type: integer
          format: int32
          example: 0
        district:
          type: string
          example: DownTown
        street:
          type: string
          example: Broadway
        homeNum:
          type: string
          example: 123
        shiftDirectionsPL:
          type: string
          example: Z lewej strony budynku
        shiftDirectionsEN:
          type: string
          example: Left Side of Building
        shiftDirectionsUA:
          type: string
          example: Ліва сторона будівлі
        shiftDirectionsRU:
          type: string
          example: Левая сторона здания
...
    LocalTime:
      type: object
      properties:
        hour:
          type: integer
          format: int32
        minute:
          type: integer
          format: int32
        second:
          type: integer
          format: int32
        nano:
          type: integer
          format: int32